# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Beauty Parlor API",
        "version": "v1"
    },
    "paths": {
...
        "/api/v1/signup": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Sign up",
                "description": "Signup",
                "operationId": "Signup",
                "requestBody": {
                    "description": "Sign up",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "first_name",
                                    "last_name",
                                    "phone_number"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "maxLength": 40,
                                        "example": "daniel"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "maxLength": 40,
                                        "example": "chisom"
                                    },
                                    "username": {
                                        "type": "string",
                                        "maxLength": 40,
                                        "example": "chisom"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "test@gmail.com"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "maxLength": 40,
                                        "example": "08012345678"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "chisonum"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong while processing your request."
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "500"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "error_debug": {
                                            "type": "string",
                                            "example": "The provided credentials are incorrect."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/signin": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Sign in",
                "description": "Login with email and password",
                "operationId": "Signin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "test@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "chisonum"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "500": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Something went wrong while processing your request."
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "500"
                                        },
                                        "success": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "error_debug": {
                                            "type": "string",
                                            "example": "The provided credentials are incorrect."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}